refactor: move files and applied formatter
