+++ src/serial.js
